server/agents/debugger/requirements.txt
fastapi
uvicorn
pydantic
requests
langchain
transformers
structlog
python-dotenv
pytest